work version

+++ src/main/java/by/tms/entity/User.java
@@ -1,7 +1,7 @@
 package by.tms.entity;
 
 public class User {
-    private long id;
+    private int id;
     private String name;
     private String username;
     private String password;
@@ -15,18 +15,18 @@ public class User {
         this.password = password;
     }
 
-    public User(long id, String name, String username, String password) {
+    public User(int id, String name, String username, String password) {
         this.id = id;
         this.name = name;
         this.username = username;
         this.password = password;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
